v4.5 Admin panel design, poprawki PHP
*Design panelu admin:
-Usunięcie zakładki kontakt
-Usunięcie niepotrzebnych opcji w zakładce profil i zmiana stylistyki
-Poprawa ikon panelu
-Zmiana szaty graficznej panelu admina na bg-info (bootstrap)
-Utworzenie klas zarządzających stylistyką panelu jako "<nazwa>-admin"
-Stylistyka tabel, przycisków i układu blokowego stron: klienci, kluby i szczegóły klubu
-Zedytowane logo panelu admina zostało poprawnie rozpoznawane i dynamicznie podmieniane przy pomocy jQuery
-Pełna responsywność wyświetlamych tabel oraz formularzy

*Stylistyka panelu użytkownika
-Dodanie układu blokowego dla każdej podstrony "szegółów badań"

*Komunikaty "zapytania"
-Obsługa zapytań użytkownika/admina o potwierdzenie wykonania danej akcji
-Wykorzystana sesja PHP oraz połączenia z obiektem DOM w JS
-Zarządzanie w plikach daddy_im_shy.php oraz ownJs.js
-Stylizacja i pełna responsywność wyświetlanych komunikatów

*Poprawki PHP
-Tabele klientów oraz klubów wyświetlają komunikat gdy baza jest pusta (zamiast pustej tablicy)
-Kilka przycisków zarządzających przejściami do odpowiednich podstron zostały zastąpione dropboxem
-Zabezpieczenie w przypadku braku podanych parametrów oraz zamykanie połączenia z bazą
-Usuwanie sesji admina w pliku wyloguj.php
-Wpisane wartości w wyszukiwarkę klienta/klubu zostają w oknie
-Wybranie pustego formularza wyszukiwarki wyświetla wszystkie rekordy [wyświetlało pustą stronę]
-rozchodniaczki usun_klub.php oraz usun_klienta.php przyjmują parametry metodą GET w celu przystępniejszego przekazywania danych

+++ rozchodniaczki/usun_klub.php
@@ -1,19 +1,30 @@
 <?php
     session_start();
-	if (session_status() == PHP_SESSION_NONE) {
-		header('Location: ../../logowanie.php');
+	
+	 function jump_to_page($mode,$top,$bottom) {
+        header('Location: ../panel_admina.php');
+		$_SESSION['error'] = 'loadToast(\''.$mode.'\',\''.$top.'\',\''.$bottom.'\')';
+    }
+
+	if(isset($_GET['id_klubu'])) {
+
+		$id_klubu = $_GET['id_klubu'];
+
+		require_once 'connect.php';
+
+		$connection = @new mysqli($host, $db_user, $db_password, $db_name);
+
+		if ($connection->connect_errno == 0) {
+			$sql = "DELETE FROM `klub` WHERE id_klubu = '$id_klubu'";
+
+			$result = @$connection->query($sql);
+			if($result){jump_to_page('0','Usunięcie klubu zakończone powodzeniem','');}
+			else{jump_to_page('3','Błąd bazy danych', 'Niepowodzenie w wykonaniu zapytania sql<br/>Command: DELETE');}
+		}
+		$connection->close();
 	}
-
-    $id_klubu = $_POST['id_klubu'];
-
-	require_once 'connect.php';
-
-	$connection = @new mysqli($host, $db_user, $db_password, $db_name);
-
-	if ($connection->connect_errno == 0) {
-        $sql = "DELETE FROM `klub` WHERE id_klubu = '$id_klubu'";
-
-        $result = @$connection->query($sql);
+	else {
+		//Brak parametrów GET
+		jump_to_page('3','Błąd logiczny','Nie podano wszystkich parametrów wymaganych do przeprowadzenia operacji!');
 	}
-	header('Location: ../panel_admina.php');
 ?>
